mini_clone/requirements.txt
# Mini Clone Requirements

# OpenAI API
openai>=1.3.0

# Environment variables
python-dotenv>=1.0.0
